Create basmati project. Borrow heavily from omnium for struct.
Has most of the basics in place, like basmati command, logging, shell.
Took useful parts from experimental and put them in hydrosheds,
plotting, and utils.
Designed to be used from outside of the project path, i.e. in new
project dir. This will be setup with a basmati.conf file which can be
used to download the hydrobasins/hydrosheds data.

Needs:
* docs
* testing
* general raster handling for iris/atmospheric data
* demo project